requirements.txt
pyyaml==6.0.1
requests==2.31.0
rich==13.7.0
groq==1.1.2
python-dotenv==1.0.0